Deleting unnecesary comments. Creating page module. Deleting styles from modules folder and adding it to index.css file. Adding Usage field to Readme.md
