--- modules/Styles/book-list.css
@@ -1,27 +1,0 @@
-.books {
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.list-title {
-  display: flex;
-  justify-content: center;
-  padding-bottom: 1rem;
-}
-
-.list-item {
-  display: flex;
-  justify-content: space-between;
-  padding-bottom: 5px;
-}
-
-.list-item:nth-child(odd) {
-  background-color: #eee;
-}
-
-.title-container {
-  display: flex;
-  gap: 0.4rem;
-}
