--- modules/Styles/config.css
@@ -1,15 +1,0 @@
-
-@import url('https://fonts.googleapis.com/css2?family=Kalam&display=swap');
-
-* {
-  margin: 0;
-  padding: 0;
-  list-style: nose;
-  text-decoration: none;
-  list-style-type: none;
-  font-family: kalam, monospace;
-}
-
-body {
-  margin: 2rem;
-}
