+++ modules/date.js
@@ -1,0 +1,6 @@
+
+import { DateTime } from './../node_modules/luxon/src/luxon.js';
+
+const currentTime = DateTime.now().setZone('America/Argentina/Buenos_Aires');
+
+export default currentTime;
